fix usages of renamed Protocol dependency ref: https://github.com/CloudburstMC/Protocol/commit/bddf7b942a57aedd2885440626ed7b68c3131c94
